backend/pyproject.toml
[project]
name = "mirofish-backend"
version = "0.1.0"
description = "MiroFish - A Simple and Universal Swarm Intelligence Engine, Predicting Anything"
requires-python = ">=3.11"
license = { text = "AGPL-3.0" }
authors = [
    { name = "MiroFish Team" }
]

dependencies = [
    # Core framework
    "flask>=3.0.0",
    "flask-cors>=6.0.0",
    
    # LLM
    "openai>=1.0.0",
    
    # Zep Cloud (when GRAPH_BACKEND=zep)
    "zep-cloud==3.13.0",
    
    # Neo4j (when GRAPH_BACKEND=neo4j)
    "neo4j>=5.0.0",
    
    # OASIS social simulation
    "camel-oasis==0.2.5",
    "camel-ai==0.2.78",
    
    # File processing
    "PyMuPDF>=1.24.0",
    # Encoding detection (non-UTF-8 text files)
    "charset-normalizer>=3.0.0",
    "chardet>=5.0.0",
    
    # Utilities
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pipreqs>=0.5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]
